Hotfix/componentizacao: componentes de Meta tags sociais, footer e botões de whatsapp criados
