+++ src/components/styles/features.scss
@@ -1,0 +1,18 @@
+#whatsapp {
+  background-color: #26CC64;
+  position: fixed;
+  bottom: 10px;
+  left: 10px;
+  width: 60px;
+  height: 60px;
+  border-radius: 100%;
+  border: none;
+  box-shadow: -1px 2px 15px #aaa;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  svg {
+    color: white;
+  }
+}
