+++ src/components/styles/main.css
@@ -1,0 +1,89 @@
+footer {
+  height: 100%;
+  background-color: black;
+  padding: 40px 0;
+}
+footer div {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 75%;
+  margin: auto;
+}
+footer div h2 {
+  color: white;
+  font-size: 24px;
+}
+footer div p {
+  text-align: center;
+  line-height: 1.2em;
+  color: white;
+}
+footer div a {
+  display: grid;
+  grid-template: auto/10% auto;
+  grid-gap: 15px;
+  justify-content: center;
+  align-items: center;
+  padding: 10px;
+  margin: 30px 0;
+  width: 75%;
+  height: 60px;
+  background-color: #26CC64;
+  color: white;
+  border-radius: 15px;
+  box-shadow: 0px 8px 1px green;
+  text-transform: uppercase;
+  font-weight: bold;
+}
+footer div a span {
+  font-size: 0.8em;
+  margin-left: 10px;
+  text-align: left;
+}
+footer div .politicas {
+  display: grid;
+  grid-template: auto/33% 33% 33%;
+  grid-gap: 5px;
+  justify-content: space-around;
+  align-items: center;
+  height: 50px;
+  width: 100%;
+  margin: 10px 0;
+}
+footer div .politicas p {
+  display: flex;
+  justify-content: center;
+  color: white;
+  height: 30px;
+  width: 100%;
+  text-align: center;
+}
+footer div .politicas p:nth-child(n+2) {
+  border-left: 1px solid;
+}
+footer div .politicas p a {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: lightskyblue;
+  font-size: 0.7em;
+  background-color: transparent;
+  border: none;
+  box-shadow: none;
+  text-transform: none;
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
+footer div .obsFacebook {
+  margin-top: 20px;
+  font-size: 0.7em;
+}
+
+@media screen and (min-width: 600px) {
+  footer {
+    height: 100%;
+  }
+}/*# sourceMappingURL=main.css.map */
