+++ src/views/styles/CaoBemResolvido.scss
@@ -819,127 +819,10 @@ $corFundo: #ECECEB;
   }
 }
 
-footer {
-  height: 100%;
-  background-color: black;
-  padding: 40px 0;
-
-  div {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 75%;
-    margin: auto;
-
-    h2 {
-      color: white;
-      font-size: 24px;
-    }
-
-    p {
-      text-align: center;
-      line-height: 1.2em;
-      color: white;
-    }
-
-    a {
-      display: grid;
-      grid-template: auto / 10% auto;
-      grid-gap: 15px;
-      justify-content: center;
-      align-items: center;
-      padding: 10px;
-      margin: 30px 0;
-      width: 75%;
-      height: 60px;
-      background-color: #26CC64;
-      color: white;
-      border-radius: 15px;
-      box-shadow: 0px 8px 1px darken(#0f0, 25%);
-      text-transform: uppercase;
-      font-weight: bold;
-
-      span {
-        font-size: .8em;
-        margin-left: 10px;
-        text-align: left;
-      }
-    }
-    
-    .politicas {
-      display: grid;
-      grid-template: auto / 33% 33% 33%;
-      grid-gap: 5px;
-      justify-content: space-around;
-      align-items: center;
-      height: 50px;
-      width: 100%;
-      margin: 10px 0;
-      
-      p {
-        display: flex;
-        justify-content: center;
-        color: white;
-        height: 30px;
-        width: 100%;
-        text-align: center;
-        
-        &:nth-child(n+2){
-          border-left: 1px solid;
-        }
-        
-        a {
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          color: lightskyblue;
-          font-size: .7em;
-          background-color: transparent;
-          border: none;
-          box-shadow: none;
-          text-transform: none;
-          height: 100%;  
-          margin: 0;
-          padding: 0;        
-        }
-      }
-    }
-
-    .obsFacebook {
-      margin-top: 20px;
-      font-size: .7em;
-    }
-  }
-}
-
-#whatsapp {
-  background-color: #26CC64;
-  position: fixed;
-  bottom: 10px;
-  left: 10px;
-  width: 60px;
-  height: 60px;
-  border-radius: 100%;
-  border: none;
-  box-shadow: -1px 2px 15px #aaa;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  svg {
-    color: white;
-  }
-}
-
 @media screen and (min-width: 600px){
   header, footer, section {
     width: 50%;
     margin: auto;
-  }
-
-  footer {
-    height: 100%;
   }
 
   .metodoUm, .metodoDois, .pqServe {
